requirements.txt
requests
beautifulsoup4
python-dotenv
spotipy
selenium
google-generativeai
